Добавлены инструменты для регистрации пользователей, точка регистрации доступна всем. Пользователи с корректными данными сохраняются в БД с зашифрованным паролем

+++ Recipes/task/src/recipes/RecipeController.java
@@ -2,6 +2,11 @@ package recipes;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 import javax.validation.Valid;
@@ -13,6 +18,8 @@ public class RecipeController {
     @Autowired
     RecipeRepository repository;
 
+    @Autowired
+    PasswordEncoder encoder;
 
     @GetMapping("/api/recipe/{id}")
     public Map<String, Object> getRecipe (@PathVariable Long id) {
@@ -31,9 +38,12 @@ public class RecipeController {
     }
 
     @PostMapping("/api/recipe/new")
-    public String saveRecipe(@Valid @RequestBody(required = false) Recipe recipe) {
+    public String saveRecipe(@AuthenticationPrincipal UserDetails details,
+                             @Valid @RequestBody(required = false) Recipe recipe) {
+        System.out.println(details.toString());
         if (recipe.getDirections().length > 0 && recipe.getIngredients().length > 0) {
             recipe.setDate(LocalDateTime.now());
+            recipe.setCreator(details.getUsername());
             repository.save(recipe);
             return "{\"id\": " + repository.count() + "}";
         } else {
@@ -41,24 +51,39 @@ public class RecipeController {
         }
     }
 
-    @DeleteMapping("/api/recipe/{id}")
-    public HttpStatus deleteRecipe(@PathVariable Long id) {
-        if (!repository.existsById(id)) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
-        } else {
+    @DeleteMapping("/api/recipe/{id}") // стоит пересмотреть условный оператор
+    public void deleteRecipe(@AuthenticationPrincipal UserDetails details, @PathVariable Long id) {
+        System.out.println(details.toString());
+        if (repository.existsById(id)
+                && details.getUsername().compareTo(repository.findById(id).get().getCreator()) == 0) {
             repository.deleteById(id);
             throw new ResponseStatusException(HttpStatus.NO_CONTENT);
+        } else {
+            if (!repository.existsById(id)) {
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+            } else if (!details.getUsername().equals(repository.findById(id).get().getCreator())) {
+                throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+            }
         }
     }
 
+
     @PutMapping("/api/recipe/{id}")
-    public HttpStatus putRecipe(@Valid @RequestBody Recipe recipe, @PathVariable Long id) {
-        if (!repository.existsById(id)) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
-        } else {
+    public void putRecipe(@AuthenticationPrincipal UserDetails details,
+                                @Valid @RequestBody Recipe recipe, @PathVariable Long id) {
+        System.out.println(details.toString());
+        if (repository.existsById(id)
+                && details.getUsername().compareTo(repository.findById(id).get().getCreator()) == 0) {
+            recipe.setCreator(details.getUsername());
             repository.updateRecipe(recipe, id);
             repository.save(recipe);
             throw new ResponseStatusException(HttpStatus.NO_CONTENT);
+        } else {
+            if (!repository.existsById(id)) {
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+            } else if (!details.getUsername().equals(repository.findById(id).get().getCreator())) {
+                throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+            }
         }
     }
